FE Beam Waveguide Propagation Solver
This C++ program is designed to solve the propagation of information within a beam waveguide using finite element methods. It calculates the eigenmodes of the waveguide by solving a generalized eigenvalue problem. The code implements the inverse power iteration algorithm to find the eigenvalues and eigenvectors associated with the waveguide. This program reads the mesh from the "unit_circle.txt" file which should be in the same directory, to read other files you need to change the waveguide.cpp. Please ensure the formatting of the file is constant.